backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
supabase
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
